# Remove incorrect kernel optimization.