Not encountering a barrier inside the first block does not guarantee that we won't see a barrier in the following blocks. The barrier semantics only require the threads within a block to reach the same barriers.

The included test showcases a simplistic example of a kernel that does satisfy the standard barrier semantics but crashes with the current HIP-CPU implementation.